pinの作成をフィアル分割
