+++ model/template.go
@@ -18,12 +18,6 @@ type TemplatePins struct {
 	Pin      uuid.UUID `db:"pin"`
 }
 
-type Pin struct {
-	Id uuid.UUID `db:"id" json:"id"`
-	X  float32   `db:"x" json:"x"`
-	Y  float32   `db:"y" json:"y"`
-}
-
 func CreateTemplate(ctx context.Context, name string, image string, pins []Pin, createdby uuid.UUID) (*uuid.UUID, error) {
 	templateID := uuid.New()
 	query := "INSERT INTO template (id, name, image, createdBy) VALUES (?, ?, ?, ?)"
@@ -31,18 +25,9 @@ func CreateTemplate(ctx context.Context, name string, image string, pins []Pin, 
 	if err != nil {
 		return nil, err
 	}
-	for _, v := range pins {
-		pinid := uuid.New()
-		queryPin := "INSERT INTO pin (id, x, y) VALUES (?, ?, ?)"
-		_, err := db.ExecContext(ctx, queryPin, pinid, v.X, v.Y)
-		if err != nil {
-			return nil, err
-		}
-		queryTeamplatePin := "INSERT INTO template_pins (template, pin) VALUES (?, ?)"
-		_, err = db.ExecContext(ctx, queryTeamplatePin, templateID, pinid)
-		if err != nil {
-			return nil, err
-		}
+	err = CreateTemplatePins(ctx, templateID, pins)
+	if err != nil {
+		return nil, err
 	}
 	return &templateID, nil
 }
